Feat: DefaultTurret 스크립트 생성후 CamoTurretLV1.cs에 상속
