+++ Assets/Prefabs/Tower/ControlUnit.prefab
@@ -139,8 +139,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5c355b3b635e540c3997a89d22e91fb1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  MaxPower: 100
-  CurrentPower: 100
+  maxPower: 100
+  currentPower: 100
+  maxHealth: 100
+  curHealth: 100
+  OnCUHpChange:
+    m_PersistentCalls:
+      m_Calls: []
 --- !u!212 &8410638605953615744
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -308,8 +313,8 @@ GameObject:
   m_Component:
   - component: {fileID: 1744049009949569818}
   - component: {fileID: 4337899493936173719}
-  - component: {fileID: 9205295638882949586}
   - component: {fileID: 469274601649993134}
+  - component: {fileID: 8477086161060408526}
   m_Layer: 0
   m_Name: STATION1
   m_TagString: Untagged
@@ -384,7 +389,34 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!60 &9205295638882949586
+--- !u!50 &469274601649993134
+Rigidbody2D:
+  serializedVersion: 4
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1150638677136147710}
+  m_BodyType: 1
+  m_Simulated: 1
+  m_UseFullKinematicContacts: 0
+  m_UseAutoMass: 0
+  m_Mass: 1
+  m_LinearDrag: 0
+  m_AngularDrag: 0.05
+  m_GravityScale: 1
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_Interpolate: 0
+  m_SleepingMode: 1
+  m_CollisionDetection: 0
+  m_Constraints: 0
+--- !u!60 &8477086161060408526
 PolygonCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -428,61 +460,25 @@ PolygonCollider2D:
   m_AutoTiling: 0
   m_Points:
     m_Paths:
-    - - {x: 13.74482, y: 12.200115}
-      - {x: 12.573322, y: 14.954807}
-      - {x: 6.860859, y: 19.407871}
-      - {x: 3.8537521, y: 19.714241}
-      - {x: -4.708412, y: 19.578823}
-      - {x: -8.672165, y: 18.851913}
-      - {x: -15.539659, y: 15.1399}
-      - {x: -18.041418, y: 10.3156595}
-      - {x: -19.68431, y: 3.94423}
-      - {x: -19.763124, y: -4.2253647}
-      - {x: -18.707754, y: -5.2034225}
-      - {x: -16.215513, y: -8.777985}
-      - {x: -11.004092, y: -14.309468}
-      - {x: -4.87075, y: -19.115604}
-      - {x: 4.971433, y: -19.35506}
-      - {x: 7.114849, y: -16.279877}
-      - {x: 14.148834, y: -18.132027}
-      - {x: 16.5262, y: -16.070518}
-      - {x: 18.314613, y: -13.973373}
-      - {x: 18.559166, y: -8.078836}
-      - {x: 18.647503, y: -6.8366394}
-      - {x: 18.820255, y: -4.9236755}
-      - {x: 19.747189, y: -4.094307}
-      - {x: 19.900936, y: 0.19813919}
-      - {x: 19.834904, y: 3.4739609}
-      - {x: 19.182827, y: 7.1908264}
-      - {x: 17.559155, y: 10.01902}
+    - - {x: 6.8423157, y: 19.223923}
+      - {x: 4.6461945, y: 19.787102}
+      - {x: -4.7809067, y: 19.655384}
+      - {x: -8.540871, y: 18.749126}
+      - {x: -13.840668, y: 16.662128}
+      - {x: -15.404121, y: 15.134953}
+      - {x: -18.265125, y: 10.210454}
+      - {x: -18.984222, y: 7.163559}
+      - {x: -19.773338, y: 4.016014}
+      - {x: -19.85257, y: -4.2548065}
+      - {x: -16.539295, y: -8.943418}
+      - {x: -10.062004, y: -15.361076}
+      - {x: -4.768543, y: -19.559498}
+      - {x: 4.90625, y: -19.359947}
+      - {x: 11.256891, y: -17.210453}
+      - {x: 17.450161, y: -10.873369}
+      - {x: 19.917099, y: -4.985634}
+      - {x: 19.367733, y: 7.0696483}
   m_UseDelaunayMesh: 0
---- !u!50 &469274601649993134
-Rigidbody2D:
-  serializedVersion: 4
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1150638677136147710}
-  m_BodyType: 1
-  m_Simulated: 1
-  m_UseFullKinematicContacts: 0
-  m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 0
-  m_AngularDrag: 0.05
-  m_GravityScale: 1
-  m_Material: {fileID: 0}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_Interpolate: 0
-  m_SleepingMode: 1
-  m_CollisionDetection: 0
-  m_Constraints: 0
 --- !u!1 &2135681642179865673
 GameObject:
   m_ObjectHideFlags: 0
