+++ Assets/Scripts/Tower/ActivateTower.cs
@@ -1,0 +1,8 @@
+namespace Tower
+{
+    interface ActivateTower
+    {
+        void ActivateTurret();
+        void DeactivateTurret();
+    }
+}
