+++ Assets/Scripts/Tower/CamoTurret LV1.cs.meta
@@ -1,11 +1,3 @@
 fileFormatVersion: 2
-guid: 463467f4a02be42a38d700f2b72797bb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+guid: b1bd8bb9366149bab62297b29550d748
+timeCreated: 1730696381
