+++ Assets/Scripts/Tower/DefaultTurret.cs.meta
@@ -1,11 +1,3 @@
 fileFormatVersion: 2
-guid: 6f6f4029f42d84861b30084148074fc7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+guid: 2e196071f7c64457bfb7a2462ab62a98
+timeCreated: 1730696421
